Move game source files and adjust imports
The game source files (TextureManager.cpp, Handler.cpp, GameObject.cpp, Game.cpp) have been moved into a new directory, GameHeaders/SourceFiles/. This is reflected in the updates to the import statements in these files and in CMakeLists.txt. A new game object, Background, was also created with its Update() and Render() methods, and was added to the main game loop as part of this commit. FIX THE BACKGROUND

Signed-off-by: DomAtTheShack <dominic@domsmacshack.com>