Add firing mechanism to PlayerBase and Cursor classes
Enhancements were made to the PlayerBase and Cursor classes to support a new firing mechanism. The 'fire' function in PlayerBase has been modified to take coordinates as parameters. A 'HandleInput' function has also been added to both classes to handle the fire event. The 'fired' boolean indicates if the object has already fired.